requirements.txt
amqp==2.4.2
atomic==0.7.3
Babel==2.6.0
billiard==3.6.0.0
celery==4.3.0
certifi==2018.11.29
cffi==1.12.3
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
dataclasses==0.6
defusedxml==0.5.0
Django==2.0.7
django-allauth==0.38.0
django-cors-headers==2.4.0
django-countries==5.3.2
django-filter==2.1.0
django-hashers-passlib==0.4
django-phonenumber-field==2.2.0
django-rest-auth==0.9.3
django-rest-swagger==2.2.0
django-stubs==0.9.0
djangorestframework==3.9.1
djangorestframework-stubs==0.1.0
drf-nested-routers==0.91
et-xmlfile==1.0.1
idna==2.8
isort==4.3.4
itypes==1.1.0
jdcal==1.4.1
Jinja2==2.10
kombu==4.5.0
MarkupSafe==1.1.0
mypy==0.670
mypy-extensions==0.4.1
numpy==1.16.4
oauthlib==3.0.1
openapi-codec==1.3.2
openpyxl==2.6.2
pandas==0.24.2
passlib==1.7.1
phonenumbers==8.10.4
Pillow==5.4.1
pkg-resources==0.0.0
psycopg2==2.7.7
psycopg2-binary==2.7.7
PxPayPy==0.1.3a1
pycparser==2.19
PyMySQL==0.9.3
python-dateutil==2.8.0
python-decouple==3.1
python-mimeparse==1.6.0
python3-openid==3.1.0
pytz==2018.9
redis==3.2.1
requests==2.21.0
requests-oauthlib==1.2.0
RestAuthCommon==0.7.0
simplejson==3.16.0
six==1.12.0
South==1.0.2
stripe==2.21.0
typed-ast==1.3.1
typing-extensions==3.7.2
uritemplate==3.0.0
urllib3==1.24.1
uuid==1.30
vine==1.3.0
xlrd==1.2.0
XlsxWriter==1.1.8
